- Updating Talkers compilation for deployment, but there is still more work to do - Talkers Test cycle is complete

+++ Talkers/Utils/Random.hpp
@@ -8,14 +8,14 @@
 #ifndef Random_h
 #define Random_h
 
-// https://stackoverflow.com/questions/6942273/how-to-get-a-random-element-from-a-c-container
+// thx https://stackoverflow.com/questions/6942273/how-to-get-a-random-element-from-a-c-container
 
 #include  <random>
 #include  <iterator>
 
 template<typename Iter, typename RandomGenerator>
 Iter select_randomly(Iter start, Iter end, RandomGenerator& g) {
-	std::uniform_int_distribution<> dis(0, std::distance(start, end) - 1);
+	std::uniform_int_distribution<> dis(0, (int)std::distance(start, end) - 1);
 	std::advance(start, dis(g));
 	return start;
 }
